+++ ufcg/src/main/java/br/com/askufcg/dtos/PostCommentDTO.java
@@ -1,4 +1,4 @@
-package br.com.askufcg.dto;
+package br.com.askufcg.dtos;
 
 import lombok.Data;
