+++ ufcg/src/main/java/br/com/askufcg/repositories/AnswerRepository.java
@@ -1,4 +1,4 @@
-package br.com.askufcg.repository;
+package br.com.askufcg.repositories;
 
 import br.com.askufcg.models.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
