+++ ufcg/src/main/java/br/com/askufcg/repositories/CommentRepository.java
@@ -1,4 +1,4 @@
-package br.com.askufcg.repository;
+package br.com.askufcg.repositories;
 
 import br.com.askufcg.models.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
