+++ ufcg/src/main/java/br/com/askufcg/repositories/QuestionRepository.java
@@ -1,4 +1,4 @@
-package br.com.askufcg.repository;
+package br.com.askufcg.repositories;
 
 import br.com.askufcg.models.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
